Implemented opengl shader
