+++ rust/sapfire/crates/sapfire_renderer/src/opengl_wrapper/mod.rs
@@ -1,0 +1,2 @@
+mod opengl_shader;
+pub use opengl_shader::*;
