+++ rust/sapfire/src/main.rs
@@ -1,4 +1,4 @@
-use sapfire_renderer::renderer::*;
+use sapfire_renderer::*;
 fn main() {
     Renderer::new(RenderingAPI::OpenGL).run();
 }
