+++ rust/sapfire/triangle.frag
@@ -1,0 +1,6 @@
+#version 330 core
+  out vec4 final_color;
+
+  void main() {
+    final_color = vec4(1.0, 1.0, 1.0, 1.0);
+  }
