+++ rust/sapfire/triangle.vert
@@ -1,0 +1,6 @@
+#version 330 core
+layout (location = 0) in vec3 pos;
+
+void main() {
+    gl_Position = vec4(pos.x, pos.y, pos.z, 1.0);
+    }
